scraper/pyproject.toml
[tool.poetry]
name = "athletics-ni-scraper"
version = "0.1.0"
description = "PDF downloader and parser for athletics NI website"
authors = ["Martin McKeaveney <martinmckeaveney@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
requests = "^2.21"
tabula-py = "^1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
